# backend/openapi.yaml
openapi: 3.0.0
info:
  title: GamePlatform API
  version: v1.0.0
  description: API for managing a gaming platform including championships, events, user registrations, results, and more.
servers:
  - url: /api # Assuming all backend routes are prefixed with /api
components:
  schemas:
...
    FileMetadata:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        uploader_user_id:
          type: string
          format: uuid
          readOnly: true
        storage_bucket:
          type: string
        storage_path:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        upload_status:
          type: string
          enum: [pending, completed, failed]
        entity_type:
          type: string
          nullable: true
        entity_id_int:
          type: integer
          nullable: true
        entity_id_uuid:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        public_url: # Added by frontend/backend when listing
          type: string
          format: url
          nullable: true
          readOnly: true
    SignedUploadURLRequest:
      type: object
      properties:
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        bucketName:
          type: string
          enum: [gallery, avatars, track_layouts, general_uploads] # Example buckets
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        entity_type:
          type: string
          nullable: true
        entity_id_int:
          type: integer
          nullable: true
        entity_id_uuid:
          type: string
          format: uuid
          nullable: true
      required:
        - fileName
        - fileType
        - fileSize
        - bucketName
    SignedUploadURLResponse:
      type: object
      properties:
        signedUrl:
          type: string
          format: url
        storagePath:
          type: string
        fileRecordId:
          type: string
          format: uuid
    UploadCompleteRequest:
      type: object
      properties:
        fileRecordId:
          type: string
          format: uuid
        success:
          type: boolean
        errorMessage:
          type: string
          nullable: true
      required:
        - fileRecordId
        - success

    # --- Comment Schemas (Example - User can expand) ---
...
  securitySchemes:
    bearerAuth: # Can be named anything
      type: http
      scheme: bearer
      bearerFormat: JWT # Supabase uses JWT
security:
  - bearerAuth: [] # Applies Bearer token auth globally to all operations by default. Override locally if needed.
paths:
...
  /files/signed-upload-url:
    post:
      summary: Get a signed URL for file upload
      tags: [Files]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignedUploadURLRequest'
      responses:
        '200':
          description: Signed URL and file record ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUploadURLResponse'
        '400':
          description: Invalid input (e.g., file size, type, bucket).
        '401':
          description: Unauthorized.
  /files/upload-complete:
    post:
      summary: Notify backend of upload completion status
      tags: [Files]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCompleteRequest'
      responses:
        '200':
          description: Upload status updated.
        '400':
          description: Invalid input.
        '403':
          description: Forbidden (not uploader or admin).
        '404':
          description: File record not found.
  /files:
    get:
      summary: List files with filters
      tags: [Files]
      # Security can be optional if some files are public via RLS on buckets/tables
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: entity_type
          in: query
          schema: { type: string }
        - name: entity_id_int
          in: query
          schema: { type: integer }
        # ... other filters like entity_id_uuid, uploader_user_id, bucketName
      responses:
        '200':
          description: A list of file metadata with public URLs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileMetadata'
                  total: { type: integer }
                  # ... pagination fields
  /files/{fileId}:
    delete:
      summary: Delete a file and its metadata (Owner or Admin)
      tags: [Files]
      security:
        - bearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: File deleted successfully.
        '403':
          description: Forbidden.
        '404':
          description: File record not found.
  /files/{fileId}/metadata:
    put:
      summary: Update file metadata (Owner or Admin)
      tags: [Files]
      security:
        - bearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Metadata updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
        '400':
          description: Invalid input.
        '403':
          description: Forbidden.
        '404':
          description: File record not found.

  # --- Comment Routes (Example - User can expand) ---
...